Move Host entity to entities package
