+++ app-shared/src/commonMain/kotlin/org/pointyware/artes/data/hosts/SqlDelightHostDao.kt
@@ -2,7 +2,7 @@ package org.pointyware.artes.data.hosts
 
 import org.pointyware.artes.data.hosts.db.HostDb
 import org.pointyware.artes.data.hosts.db.Hosts
-import org.pointyware.artes.entities.hosts.Host
+import org.pointyware.artes.entities.Host
 import org.pointyware.artes.services.openai.OpenAi
 
 /**
