+++ core-entities/src/commonMain/kotlin/org/pointyware/artes/entities/Agent.kt
@@ -1,7 +1,5 @@
 package org.pointyware.artes.entities
 
-import org.pointyware.artes.entities.hosts.Host
-
 /**
  *
  */
