+++ core-entities/src/commonMain/kotlin/org/pointyware/artes/entities/Host.kt
@@ -1,4 +1,4 @@
-package org.pointyware.artes.entities.hosts
+package org.pointyware.artes.entities
 
 /**
  *
