+++ services/open-ai/src/commonMain/kotlin/org/pointyware/artes/services/openai/OpenAi.kt
@@ -1,6 +1,6 @@
 package org.pointyware.artes.services.openai
 
-import org.pointyware.artes.entities.hosts.Host
+import org.pointyware.artes.entities.Host
 
 /**
  *
